Add Accept-Encoding: gzip header to requests
This will give a big performance boost because the network traffic will
be much smaller.

Accoding to https://pkg.go.dev/net/http

DisableCompression, if true, prevents the Transport from
requesting compression with an "Accept-Encoding: gzip"
request header when the Request contains no existing
Accept-Encoding value. If the Transport requests gzip on
its own and gets a gzipped response, it's transparently
decoded in the Response.Body. However, if the user
explicitly requested gzip it is not automatically uncompressed.

So, no other changes should be necessary. Decompression should happen
transparently.