Add digest algorithm configuration & BLAKE3 suppport (#129)
* nothing works, but it will come

* fix: verify command

* fix: verify command

* fix: resp -> record header

* fix: verifyWARCVersion

* fix: resolve issue with local dedupe

This ensures that the prefix is consistent throughout the hash table. This will allow any hash type to work with local dedupe.

* Resolve issue with testFileSingleHashCheck

typo in record.Content vs resp.Body

* fix: remove hash check from revisit options that can't use it

* cleaner tests

* refactor GetDigest

* move tests to digest_test.go

* add other possible empty digests

* duh

* feat: add tests to validate empty byte hashes

* fix: blake3 empty payload hash based on new tests

* fix: ensure that only SHA1 can be used on Doppelganger and CDX for now

Doppelganger will be modified later to support anything, but CDX (as far as IA CDX) is concerned will not be modified.

---------

Co-authored-by: Jake L <NGTmeaty@users.noreply.github.com>